Improve [Obsolete] warnings - include the Nlog version when it became obsolete (#1814)
* Document the version an elemement became obsolete

Document on XML comment as well as the [Obsolete] attribute the NLog
version where a class, method or property was marked as obsolete.

This might help in the future releases to evaluate if an element should
be removed completely from the code and provide the users with more
information.

* Versions updated when elements marked as obsolete

Updated the XML documentation and the obsolete messages with the actual
versions that certain elements marked as obsolete.